requirements.txt
amazon-transcribe==0.6.2
aiofile==3.8.8
boto3==1.35.19
ffprobe==0.5
pydub==0.25.1
ffmpeg==1.4
wave==0.0.2
google-cloud-texttospeech==2.17.2
google-cloud-speech==2.27.0
playsound3==2.2.2
pydantic==2.9.2
pyyaml==6.0.2
pvporcupine==3.0.3
pydub==0.25.1
google-cloud-vision==3.7.4
google-cloud-videointelligence==2.13.5
imageio[ffmpeg]==2.35.1
Flask==3.0.3